Changes to be committed: 	modified:   models/KondoSU2xU1.h 	modified:   models/KondoU1xU1.h 	-fehlende includes eingefuegt. Kompatibilitaet mit boost 1.66 hergestellt.

+++ models/KondoSU2xU1.h
@@ -1,6 +1,8 @@
 #ifndef KONDOMODELSU2XU1_H_
 #define KONDOMODELSU2XU1_H_
 
+#include "ParamHandler.h" // from HELPERS
+
 #include "bases/spins/BaseSU2xU1.h"
 #include "bases/fermions/BaseSU2xU1.h"
 #include "Mpo.h"
@@ -134,12 +136,13 @@ validate (qType qnum) const
 {
 	frac S_elec(qnum[1],2); //electrons have spin 1/2
 	frac Smax = S_elec;
-	for (size_t l=0; l<N_sites; ++l) { Smax+=B[l].orbitals()*frac(B[l].get_D()-1,2); } //add local spins to Smax
+	for (size_t l=0; l<N_sites; ++l) { Smax+=static_cast<int>(B[l].orbitals())*frac(B[l].get_D()-1,2); } //add local spins to Smax
 	
 	frac S_tot(qnum[0]-1,2);
 	cout << S_tot << "\t" << Smax << endl;
 	if (Smax.denominator()==S_tot.denominator() and S_tot<=Smax and qnum[0]<=2*static_cast<int>(this->N_phys) and qnum[0]>0) {return true;}
 	else {return false;}
+	// return true;
 }
 
 std::string KondoSU2xU1::
